+++ R/backend-.R
@@ -212,6 +212,7 @@ base_scalar <- sql_translator(
   # Postgres - https://www.postgresql.org/docs/8.4/static/datatype-numeric.html
   # Impala - https://impala.apache.org/docs/build/html/topics/impala_bigint.html
   as.integer64 = sql_cast("BIGINT"),
+  as.blob = sql_cast("BLOB"),
 
   c = function(...) {
     c(...)
@@ -365,11 +366,6 @@ base_scalar <- sql_translator(
   str_wrap = sql_not_supported("str_wrap")
 )
 
-base_symbols <- sql_translator(
-  pi = sql("PI()"),
-  `*` = sql("*"),
-  `NULL` = sql("NULL")
-)
 sql_exp <- function(a, x) {
   a <- as.integer(a)
   if (identical(a, 1L)) {